Update README.md
The fight for Malaria has gone on for so long. It has taken so many lives and continues to harm the most vulnerable and imporverish regions of the world, yet we have the tools to win this fight. Let's commit and end Malaria in our lifetime.

The primary goal of this project is to show why we should end one of the deadliest diseases in human history, MALARIA, and to show how that is possible in this lifetime.

Hypothesis 1: Ending Malaria Stimulates Economy.
Hypothesis 2: Ending Malaria helps keep us safer and healthier.
Hypothesis 3: Ending Malaria saves lives and stops the cycle of poverty.

References:
https://www.malarianomore.org/about-us/
https://www.kaggle.com/search?q=malaria+in%3Adatasets